build: upgrade dms agent project lib version.

+++ dms_common/src/transfer/ContainerInspectInfo.groovy
@@ -15,7 +15,7 @@ class ContainerInspectInfo {
     String logPath
     String id
 
-    Integer sizeRootFs
+    Long sizeRootFs
     String imageId
     String name
     Integer restartCount
